+++ .github/workflows/canary.yml
@@ -38,6 +38,8 @@ jobs:
       - name: Check if new canary release is required
         id: canary_required
         run: node ./tools/scripts/canary/canary_required.mjs """${{ steps.find_comment.outputs.comment-body }}""" >> $GITHUB_OUTPUT && cat $GITHUB_OUTPUT
+      - name: Prepare package names
+        run: node ./tools/scripts/canary/canary_name_change.mjs
       - name: Publish canary release
         if: steps.canary_required.outputs.skipCanary == ''
         run: pnpm -r publish --dry-run --no-git-checks
